Reorganize the structure of the application
Fourguys.py has been divided into five different
python script files, each of which contains parts
from the original file and grouped by their
respective functionalities. Furthermore, all of
them are now located within a directory called app
(or more precisely app package) which is the core
of the whole web application.

 Changes to be committed:
	deleted:    Fourguys.py
	new file:   app/__init__.py
	new file:   app/forms.py
	new file:   app/managers.py
	new file:   app/models.py
	new file:   app/views.py
