+++ app/forms.py
@@ -1,0 +1,20 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, BooleanField, IntegerField
+from wtforms.validators import InputRequired, Email, Length
+
+class signup(FlaskForm):
+    firstname = StringField('First name', validators=[InputRequired()])
+    lastname = StringField('Last name', validators=[InputRequired()])
+    email = StringField('E-mail', validators=[InputRequired(), Email(message='Invalid email'), Length(max=50)])
+    password = PasswordField('Password', validators=[InputRequired(), Length(min=8, max=80)])
+    conpassword = PasswordField('Confirm Password', validators=[InputRequired(), Length(min=8, max=80), ])
+    address = StringField('Address', validators=[InputRequired()])
+    card = PasswordField('Card Number', validators=[InputRequired()])
+
+class login1(FlaskForm):
+    email = StringField('E-mail', validators=[InputRequired(), Email(message='Invalid email'), Length(max=50)])
+    password = PasswordField('Password', validators=[InputRequired(), Length(min=8, max=80)])
+    remember = BooleanField('Remember me')
+
+class menu1(FlaskForm):
+    qty = IntegerField('qty', validators=[InputRequired()], default=0)
